sync done!

+++ src/model/remoteSave/sync.ts
@@ -4,17 +4,25 @@ import {Database} from '@nozbe/watermelondb';
 import {req} from '../../common/assistant/api';
 import {TToken} from '../../redux/types';
 import {SyncPullResult} from '../types';
-import {ChangesEvents, deleteImagesFromCache, syncPullAdapter, syncPushAdapter} from '../assist';
+import {ChangesEvents, syncPullAdapter, syncPushAdapter} from '../assist';
 import {ChaptersTableName, DiaryTableName, NotesTableName, PagesTableName, PhotosTableName} from '../schema';
-import {database} from '../../AppContainer';
+import {deletePhotosFromS3_URI, downloadNewPhotosS3, uploadNewPhotosOnS3} from './s3Bucket';
+import {SyncActions} from '../../components/menu/ModalSaveData';
 
 
-export async function syncDB(database: Database, token: TToken | null, userId: number | null, deletedPhotos: string[]) {
+export async function syncDB(
+  database: Database,
+  token: TToken | null,
+  userId: number | null,
+  deletedPhotos: string[],
+  onProgress: (total: number, processed: number, action: SyncActions)=> void
+) {
   await synchronize({
     database,
     sendCreatedAsUpdated: true,
     pullChanges: async ({lastPulledAt, schemaVersion, migration}) => {
       try {
+        console.log('pull');
         const chapters = await req(token).get<SyncPullResult>('/chapters/sync', {params: {
           lastPulledAt,
           schemaVersion,
@@ -46,18 +54,34 @@ export async function syncDB(database: Database, token: TToken | null, userId: n
           ...notes.data?.changes,
           ...photosByMonth.data?.changes,
         };
+        syncChanges[PhotosTableName].deleted = [];
         const diaryId = notes.data?.diaryId;
         const diaryIds = await database.get(DiaryTableName).query().fetchIds();
         const adaptedChanges =  syncPullAdapter({changes: syncChanges, timestamp: notes.data?.timestamp}, diaryIds, diaryId);
 
         if (photosByMonth.data?.changes?.[PhotosTableName]?.[ChangesEvents.created]?.length ||
           photosByMonth.data?.changes?.[PhotosTableName]?.[ChangesEvents.updated]?.length) {
-          await database?.write(async () => {
-            await database.get(PhotosTableName).query().destroyAllPermanently();
-          });
+          if (diaryId !== diaryIds[0])
+            await database?.write(async () => {
+              await database.get(PhotosTableName).query().destroyAllPermanently();
+            });
         }
-        //deleteImagesFromCache();
-        console.log(syncChanges);
+        const notesPhotoCreated = notes.data?.changes?.[NotesTableName]?.created || [];
+        const notesPhotoUpdated = notes.data?.changes?.[NotesTableName]?.updated || [];
+        //const notesPhotoDeleted = notes.data?.changes?.[NotesTableName]?.deleted || [];
+        const photoByMonthCreated = photosByMonth.data?.changes?.[PhotosTableName]?.created || [];
+        const photoByMonthUpdated = photosByMonth.data?.changes?.[PhotosTableName]?.updated || [];
+        console.log('photoByMonthCreated', photoByMonthCreated);
+        console.log('photoByMonthUpdated', photoByMonthUpdated);
+        await deletePhotosFromS3_URI(deletedPhotos).catch(console.log);
+        console.log('adaptedChanges', adaptedChanges);
+
+        await downloadNewPhotosS3(
+          [...notesPhotoCreated, ...notesPhotoUpdated, ...photoByMonthCreated, ...photoByMonthUpdated],
+          onProgress
+        ).catch(console.log);
+        onProgress(1, 0, SyncActions.Other);
+        //await deletePhotosFromS3([...notesPhotoDeleted], database).catch(console.log);
         return adaptedChanges;
       } catch (err) {
         console.error(err.response?.data || err.response || err);
@@ -66,6 +90,7 @@ export async function syncDB(database: Database, token: TToken | null, userId: n
     },
     pushChanges: async ({changes, lastPulledAt}) => {
       try {
+        console.log('push');
         const chapterDTO = syncPushAdapter({changes, lastPulledAt}, userId, ChaptersTableName);
         await req(token).post('/chapters/sync', chapterDTO);
 
@@ -78,8 +103,17 @@ export async function syncDB(database: Database, token: TToken | null, userId: n
         const photoDTO = syncPushAdapter({changes, lastPulledAt}, userId, PhotosTableName);
         await req(token).post('/photos-by-month/sync', photoDTO);
 
-        console.log('photoDTO', photoDTO);
-        await req(token).delete('/photos', {data: {deletedPhotos}});
+        //await req(token).delete('/photos', {data: {deletedPhotos}});
+
+        const notesPhotoCreated = changes?.[NotesTableName]?.created || [];
+        const notesPhotoUpdated = changes?.[NotesTableName]?.updated || [];
+        const photoByMonthCreated = changes?.[PhotosTableName]?.created || [];
+        const photoByMonthUpdated = changes?.[PhotosTableName]?.updated || [];
+        await uploadNewPhotosOnS3(
+          [...notesPhotoCreated, ...notesPhotoUpdated, ...photoByMonthCreated, ...photoByMonthUpdated],
+          onProgress)
+          .catch(console.log);
+
       } catch (err) {
         console.error(err.response?.data || err.response || err);
         throw new Error('error push sync');
@@ -87,6 +121,16 @@ export async function syncDB(database: Database, token: TToken | null, userId: n
     },
     migrationsEnabledAtVersion: 1,
   });
+
+  /* await synchronize({
+    database,
+    sendCreatedAsUpdated: true,
+    pullChanges: async () => {
+      return
+    },
+    pushChanges: async () => {
+    }
+  });*/
 }
 
 //todo сделать метод который будет раз в месяц получать все фотки из таблиц и сверять с кешэм, чтобы удалить ненужные из кэша
